Update README.md
Incorrect or unclear instructions:

The document mentions "Simply install Cairo and scarb." However, the link provided does not directly lead to installation instructions for both Cairo and scarb, but likely only to one. Edited to provide links for both.

"scarb" is sometimes capitalized and other times not. For consistency, it should be treated the same throughout the document.

edited one instance of scarb.toml to Scarb.toml

Command consistency:

Usage of $ before some commands and not others can confuse readers about whether to include it in the terminal. It's standard in documentation to use $ to represent the command prompt, but it should be consistently used or omitted.